Remove Python 3.8 support
Python 3.8 has reached EOL.

Tests that used to require python >3.8 and kept in a special directory
got upleveled and merged with all other tests.
