+++ aiobotocore/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '2.22.0'
+__version__ = '2.23.0'
