+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "aiobotocore"
 description = "Async client for aws services using botocore and aiohttp"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 authors = [
     { name = "Nikolay Novik", email = "nickolainovik@gmail.com" },
 ]
@@ -17,7 +17,6 @@ classifiers = [
     'Natural Language :: English',
     'License :: OSI Approved :: Apache Software License',
     'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.8',
     'Programming Language :: Python :: 3.9',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
@@ -145,7 +144,7 @@ exclude = [
 line-length = 79
 indent-width = 4
 
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 # Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
