+++ tests/botocore_tests/unit/test_tokens.py
