+++ tests/test_config.py
@@ -72,13 +72,16 @@ async def test_connector_timeout():
     config = AioConfig(
         max_pool_connections=1, connect_timeout=1, retries={'max_attempts': 0}
     )
-    async with AIOServer() as server, session.create_client(
-        's3',
-        config=config,
-        endpoint_url=server.endpoint_url,
-        aws_secret_access_key='xxx',
-        aws_access_key_id='xxx',
-    ) as s3_client:
+    async with (
+        AIOServer() as server,
+        session.create_client(
+            's3',
+            config=config,
+            endpoint_url=server.endpoint_url,
+            aws_secret_access_key='xxx',
+            aws_access_key_id='xxx',
+        ) as s3_client,
+    ):
 
         async def get_and_wait():
             await s3_client.get_object(Bucket='foo', Key='bar')
@@ -106,13 +109,16 @@ async def test_connector_timeout2():
         read_timeout=1,
         retries={'max_attempts': 0},
     )
-    async with AIOServer() as server, session.create_client(
-        's3',
-        config=config,
-        endpoint_url=server.endpoint_url,
-        aws_secret_access_key='xxx',
-        aws_access_key_id='xxx',
-    ) as s3_client:
+    async with (
+        AIOServer() as server,
+        session.create_client(
+            's3',
+            config=config,
+            endpoint_url=server.endpoint_url,
+            aws_secret_access_key='xxx',
+            aws_access_key_id='xxx',
+        ) as s3_client,
+    ):
         with pytest.raises(ReadTimeoutError):
             resp = await s3_client.get_object(Bucket='foo', Key='bar')
             await resp["Body"].read()
@@ -142,12 +148,15 @@ async def test_config_http_session_cls():
 
     config = AioConfig(http_session_cls=MyHttpSession)
     session = AioSession()
-    async with AIOServer() as server, session.create_client(
-        's3',
-        config=config,
-        endpoint_url=server.endpoint_url,
-        aws_secret_access_key='xxx',
-        aws_access_key_id='xxx',
-    ) as s3_client:
+    async with (
+        AIOServer() as server,
+        session.create_client(
+            's3',
+            config=config,
+            endpoint_url=server.endpoint_url,
+            aws_secret_access_key='xxx',
+            aws_access_key_id='xxx',
+        ) as s3_client,
+    ):
         with pytest.raises(SuccessExc):
             await s3_client.get_object(Bucket='foo', Key='bar')
