+++ tests/test_lambda.py
@@ -45,7 +45,7 @@ async def test_run_lambda(iam_client, lambda_client, aws_lambda_zip):
     role_arn = await _get_role_arn(iam_client, 'test-iam-role')
     lambda_response = await lambda_client.create_function(
         FunctionName='test-function',
-        Runtime='python3.8',
+        Runtime='python3.9',
         Role=role_arn,
         Handler='lambda_function.lambda_handler',
         Timeout=10,
